make InitializeArmExtRet private to ExtendRetract as nothing else should instantiate this class either
